+++ imas_muscle3/visualization/examples/eq_psi_map/eq_psi_map.ymmsl
@@ -22,9 +22,8 @@ model:
 settings:
   source_component.source_uri: imas:hdf5?path=/home/sebbe/projects/iter_python/IMAS-MUSCLE3/data/equilibrium_flux_map
   source_component_md.source_uri: imas:hdf5?path=/home/sebbe/projects/iter_python/Waveform-Editor/data/nice-input-dd4
-  visualization_component.plot_file_path: /home/sebbe/projects/iter_python/IMAS-MUSCLE3/test_configs/eq_psi_map/eq_psi_map.py
+  visualization_component.plot_file_path: /home/sebbe/projects/iter_python/IMAS-MUSCLE3/imas_muscle3/visualization/examples/eq_psi_map/eq_psi_map.py
   visualization_component.port: 5007
-  visualization_component.throttle_interval: 0.5
   visualization_component.keep_alive: true
 implementations:
   visualization_component:
